move emote generator to own namespace
