+++ src/nomoretrolls/Emotes/EmoteGenerator.cs
@@ -1,4 +1,4 @@
-namespace nomoretrolls.Workflows.Reactions
+namespace nomoretrolls.Emotes
 {
     internal class EmoteGenerator : IEmoteGenerator
     {
